Move RetryWithExponentialBackoff into its own partial file to obsolete it and have a file to point to, if people want to recreate it anyway.
